+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 0.1.5
 -----
 - keep some settings between refreshes, like unsupported browser toggles, and your selected compartment per account
-- added experimental feature for compartment selection ui override. 
+- added experimental compartment picker and compartment to region mapping.
 
 0.1.4
 -----